requirements.txt
python-telegram-bot==13.7
opencv-python-headless==4.5.5.64
numpy==1.21.6